+++ packages/marionette/src/index.ts
@@ -1,1 +1,10 @@
-export * from  "./marionette";
+/**
+ * @license MIT
+ * @copyright (c) 2022 Dirt Road Dev
+ * @package @skaleproject/marionette
+ *
+ * @file index.ts
+ * @author Sawyer Cutler
+*/
+
+export * from  "./contract";
